Rename elements_list folder to elementsList
