+++ app/packs/src/components/elements/research_plans/ResearchPlanDetails.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { Panel, ListGroup, ListGroupItem, ButtonToolbar, Button, Tooltip, OverlayTrigger, Tabs, Tab, Dropdown, MenuItem } from 'react-bootstrap';
 import { unionBy, findIndex } from 'lodash';
 import Immutable from 'immutable';
-import ElementCollectionLabels from 'src/components/elements_list/ElementCollectionLabels';
+import ElementCollectionLabels from 'src/components/elementsList/ElementCollectionLabels';
 import UIActions from 'src/stores/alt/actions/UIActions';
 import ElementActions from 'src/stores/alt/actions/ElementActions';
 import DetailActions from 'src/stores/alt/actions/DetailActions';
