+++ app/packs/src/components/elements/research_plans/wellplates_tab/EmbeddedWellplate.js
@@ -4,7 +4,7 @@ import { Panel, Button, ButtonGroup, Tooltip, Overlay, OverlayTrigger, Table, Gl
 import Aviator from 'aviator';
 import UIStore from 'src/stores/alt/stores/UIStore';
 import { wellplateShowOrNew } from 'src/components/routesUtils';
-import ElementCollectionLabels from 'src/components/elements_list/ElementCollectionLabels';
+import ElementCollectionLabels from 'src/components/elementsList/ElementCollectionLabels';
 import ResearchPlan from 'src/models/ResearchPlan';
 import Wellplate from 'src/models/Wellplate';
