+++ app/packs/src/components/elements/screens/ScreenDetails.js
@@ -10,7 +10,7 @@ import Immutable from 'immutable';
 import ConfirmClose from 'src/components/common/ConfirmClose';
 import DetailActions from 'src/stores/alt/actions/DetailActions';
 import ElementActions from 'src/stores/alt/actions/ElementActions';
-import ElementCollectionLabels from 'src/components/elements_list/ElementCollectionLabels';
+import ElementCollectionLabels from 'src/components/elementsList/ElementCollectionLabels';
 import ElementDetailSortTab from 'src/components/elements/ElementDetailSortTab';
 import LoadingActions from 'src/stores/alt/actions/LoadingActions';
 import PrintCodeButton from 'src/components/common/PrintCodeButton';
