+++ app/packs/src/components/elementsList/ElementsList.js
@@ -5,8 +5,8 @@ import KeyboardActions from 'src/stores/alt/actions/KeyboardActions';
 import UIActions from 'src/stores/alt/actions/UIActions';
 import UserActions from 'src/stores/alt/actions/UserActions';
 import MatrixCheck from 'src/components/common/MatrixCheck';
-import ElementsTable from 'src/components/elements_list/ElementsTable';
-import ElementsTableSettings from 'src/components/elements_list/ElementsTableSettings';
+import ElementsTable from 'src/components/elementsList/ElementsTable';
+import ElementsTableSettings from 'src/components/elementsList/ElementsTableSettings';
 import ElementStore from 'src/stores/alt/stores/ElementStore';
 import UIStore from 'src/stores/alt/stores/UIStore';
 import UserStore from 'src/stores/alt/stores/UserStore';
